+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased][unreleased]
 
 - Support `OR` in queries
+- New API methods for Database:
+  - async upsert(table, record, condition): Object
 
 ## [1.1.3][] - 2021-05-25